refactored, misc edits
refactored class names, removed unneeded classes (BaseActivity) and gave their functionality (if any) to other classes, minor cosmetic edits to the reply views
