--- app/src/main/java/com/example/daniel/heymayo/BaseActivity.java
@@ -1,40 +1,0 @@
-package com.example.daniel.heymayo;
-
-import android.app.ProgressDialog;
-import android.support.v7.app.AppCompatActivity;
-
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.iid.FirebaseInstanceId;
-
-/**
- * Created by jsayler on 10/15/17.
- */
-
-public class BaseActivity extends AppCompatActivity {
-
-    private ProgressDialog mProgressDialog;
-
-    public void showProgressDialog() {
-        if (mProgressDialog == null) {
-            mProgressDialog = new ProgressDialog(this);
-            mProgressDialog.setCancelable(false);
-            mProgressDialog.setMessage("Loading...");
-        }
-
-        mProgressDialog.show();
-    }
-
-    public void hideProgressDialog() {
-        if (mProgressDialog != null && mProgressDialog.isShowing()) {
-            mProgressDialog.dismiss();
-        }
-    }
-
-    public String getUid() {
-        return FirebaseAuth.getInstance().getCurrentUser().getUid();
-    }
-
-    public String getToken() {
-        return FirebaseInstanceId.getInstance().getToken();
-    }
-}
