+++ app/src/main/java/com/example/daniel/heymayo/RequestPostActivity.java
@@ -2,8 +2,6 @@ package com.example.daniel.heymayo;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
@@ -12,6 +10,7 @@ import android.widget.Toast;
 
 import com.example.daniel.heymayo.models.Request;
 import com.example.daniel.heymayo.models.User;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -21,7 +20,7 @@ import com.google.firebase.database.ValueEventListener;
 import java.util.HashMap;
 import java.util.Map;
 
-public class RequestPostActivity extends BaseActivity {
+public class RequestPostActivity extends MainActivity {
 
     private static final String TAG = "RequestPostActivity";
     private static final String REQUIRED = "Required";
@@ -35,7 +34,7 @@ public class RequestPostActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_view_requests);
+        setContentView(R.layout.activity_request_post);
 
         mDatabase = FirebaseDatabase.getInstance().getReference();
         mBodyField = (EditText) findViewById(R.id.field_body);
@@ -120,6 +119,10 @@ public class RequestPostActivity extends BaseActivity {
         childUpdates.put("/requests/" + key, postValues);
         childUpdates.put("/user-posts/" + userId + "/requests/" + key, postValues);
         mDatabase.updateChildren(childUpdates);
+    }
+
+    public String getUid() {
+        return FirebaseAuth.getInstance().getCurrentUser().getUid();
     }
 
     //deprecated - not needed (no menu bar)
