+++ app/src/main/java/com/example/daniel/heymayo/fragments/RequestFragment.java
@@ -10,10 +10,10 @@ import android.support.v7.widget.SnapHelper;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
-import com.example.daniel.heymayo.PostViewHolder;
 import com.example.daniel.heymayo.R;
-import com.example.daniel.heymayo.ReplyPostActivity;
+import com.example.daniel.heymayo.ReplyActivity;
 import com.example.daniel.heymayo.models.Request;
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
 import com.firebase.ui.database.FirebaseRecyclerOptions;
@@ -21,22 +21,22 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
 
-public class PostListFragment extends Fragment {
+public class RequestFragment extends Fragment {
 
-    private static final String TAG = "PostListFragment";
+    private static final String TAG = "RequestFragment";
 
     private DatabaseReference mDatabase;
     private FirebaseRecyclerAdapter<Request, PostViewHolder> mAdapter;
     private RecyclerView mRecycler;
     private LinearLayoutManager mManager;
 
-    public PostListFragment() {}
+    public RequestFragment() {}
 
     @Override
     public View onCreateView (LayoutInflater inflater, ViewGroup container,
                               Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View rootView = inflater.inflate(R.layout.fragment_all_requests, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_request, container, false);
 
         mDatabase = FirebaseDatabase.getInstance().getReference();
 
@@ -69,7 +69,7 @@ public class PostListFragment extends Fragment {
             @Override
             public PostViewHolder onCreateViewHolder(ViewGroup viewGroup, int i) {
                 LayoutInflater inflater = LayoutInflater.from(viewGroup.getContext());
-                return new PostViewHolder(inflater.inflate(R.layout.item_request, viewGroup, false));
+                return new PostViewHolder(inflater.inflate(R.layout.item_request_fragment, viewGroup, false));
             }
 
             @Override
@@ -84,8 +84,8 @@ public class PostListFragment extends Fragment {
                     @Override
                     public void onClick(View v) {
                         // Launch PostDetailActivity
-                        Intent intent = new Intent(getActivity(), ReplyPostActivity.class);
-                        intent.putExtra(ReplyPostActivity.EXTRA_POST_KEY, postKey);
+                        Intent intent = new Intent(getActivity(), ReplyActivity.class);
+                        intent.putExtra(ReplyActivity.EXTRA_POST_KEY, postKey);
                         startActivity(intent);
                     }
                 });
@@ -130,6 +130,30 @@ public class PostListFragment extends Fragment {
 
     public Query getQuery(DatabaseReference databaseReference) {
         return databaseReference.child("requests");
+    }
+
+    public class PostViewHolder extends RecyclerView.ViewHolder {
+
+        //public TextView titleView;
+        //public TextView authorView;
+        public TextView userId;
+        public TextView bodyView;
+
+        public PostViewHolder(View itemView) {
+            super(itemView);
+
+            //titleView = itemView.findViewById(R.id.post_title);
+            //authorView = itemView.findViewById(R.id.post_author);
+            userId = itemView.findViewById(R.id.post_author);
+            bodyView = itemView.findViewById(R.id.post_body);
+        }
+
+        public void bindToPost(Request request) {
+            //titleView.setText(request.title);
+            //authorView.setText(request.author);
+            userId.setText(request.uid);
+            bodyView.setText(request.body);
+        }
     }
 
 }
