+++ app/src/main/res/layout/fragment_request.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.PostListFragment">
+    tools:context=".fragments.RequestFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/messages_list"
@@ -12,6 +12,6 @@
         android:clipToPadding="false"
         android:padding="5dp"
         android:scrollbars="vertical"
-        tools:listitem="@layout/item_request" />
+        tools:listitem="@layout/item_request_fragment" />
 
 </FrameLayout>
