+++ app/src/main/res/layout/item_reply_others.xml
@@ -5,7 +5,6 @@
     android:layout_height="wrap_content"
     xmlns:card_view="http://schemas.android.com/apk/res-auto"
     android:layout_marginTop="4dp"
-    android:layout_marginStart="64dp"
     card_view:cardCornerRadius="24dp">
 
     <RelativeLayout
@@ -20,14 +19,13 @@
             android:id="@+id/top_bar">
 
             <TextView
-                android:id="@+id/me"
-                android:text="Me"
+                android:id="@+id/reply_author"
                 android:layout_width="0dp"
                 android:layout_height="24dp"
                 android:gravity="center|start"
                 android:layout_weight="10"
                 android:textStyle="bold"
-                tools:ignore="HardcodedText" />
+                tools:text="John Doe" />
 
             <ImageView
                 android:id="@+id/star_outline"
@@ -36,13 +34,12 @@
                 android:layout_gravity="center_horizontal|end"
                 android:src="@drawable/ic_toggle_star_outline_24"
                 android:contentDescription="star outline"
-                tools:ignore="HardcodedText"
-                android:visibility="invisible"/>
+                tools:ignore="HardcodedText" />
 
         </LinearLayout>
 
         <TextView
-            android:id="@+id/requestor_reply_body"
+            android:id="@+id/reply_body"
             style="@style/TextAppearance.AppCompat.Small"
             android:layout_below="@+id/top_bar"
             android:layout_width="wrap_content"
